works!! f great! will change opacity
opacity of cube changed to 0.8. It actually worked with the custom marker, it read the pattern file right, the one in the direct repositoryt of relative path without https, just name of pattern file. didn't try yet with the complicated patterns like my beautiful face or starbucks pattern but it works *wink*